src/welders_arc_schemas/requirements.txt
jsonschema
pyyaml
pytest